requirements.txt
# SPDX-License-Identifier: AGPL-3.0-or-later
# This file is part of OCR Butterfly. See LICENSE for details.

Flask==3.0.0
mlx-vlm>=0.3.11
mlx>=0.20.0
Pillow>=10.3.0
Werkzeug==3.0.1
PyMuPDF
opencv-python>=4.10.0
pywebview>=5.0
transformers>=5.1.0
